refactor(BaseExchangeApi): moving class from types
updating imports
