+++ src/api/CoinBase.ts
@@ -1,6 +1,7 @@
 import fetch, { Response } from 'node-fetch';
 import BigNumber from 'bignumber.js';
-import { BaseExchangeApi, ExchangeApi } from '../types/ExchangeApi';
+import ExchangeApi from '../types/ExchangeApi';
+import BaseExchangeApi from './BaseExchangeApi';
 
 const URL = 'https://api.coinbase.com/v2/exchange-rates';
 
@@ -25,7 +26,7 @@ const CURRENCY_MAPPING: Record<string, string> = {
     RBTC: 'RBTC'
 };
 
-export class CoinBase extends BaseExchangeApi implements ExchangeApi {
+export default class CoinBase extends BaseExchangeApi implements ExchangeApi {
     constructor() {
         super('CoinBase', CURRENCY_MAPPING);
     }
